wip: move icons to components
