+++ ui/components/icons/AppsIcon.tsx
