+++ ui/components/icons/CopyDocumentIcon.tsx
