+++ ui/components/icons/EosIcon.tsx
