+++ ui/components/icons/GearSixIcon.tsx
