+++ ui/components/icons/GlobeIcon.tsx
