+++ ui/components/icons/LightModeIcon.tsx
