+++ ui/components/icons/MetamaskIcon.tsx
