+++ ui/components/icons/PoygonIcon.tsx
