+++ ui/components/icons/SignOutIcon.tsx
