+++ ui/components/icons/SolanaIcon.tsx
