+++ ui/components/icons/SparkleIcon.tsx
