+++ ui/components/icons/UsdtIcon.tsx
