+++ ui/components/icons/WalletIcon.tsx
