+++ ui/components/icons/Web3authIcon.tsx
